requirements.txt
charset-normalizer==3.4.1
coloredlogs==15.0.1
cycler==0.11.0
einops==0.6.1
et-xmlfile==1.1.0
filelock==3.12.2
flatbuffers==25.2.10
fonttools==4.38.0
fsspec==2023.1.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.10
importlib-metadata==6.7.0
joblib==1.3.2
kiwisolver==1.4.5
markdown-it-py==2.2.0
matplotlib==3.5.1
mdurl==0.1.2
mpmath==1.3.0
numpy==1.21.6
onnx==1.12.0
onnx-simplifier==0.4.8
onnxruntime==1.12.1
opencv-python==4.11.0.86
openpyxl==3.1.3
packaging==24.0
pandas==1.3.5
pbr==6.1.1
Pillow==9.2.0
protobuf==3.19.4
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch2caffe==0.1.0
pytz==2025.1
PyYAML==6.0
requests==2.31.0
rich==13.8.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
six==1.17.0
sympy==1.10.1
testresources==2.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
timm==0.6.11
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.0
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
